PRAG-57 issue fixed : ExpressionResult.asInt() crashing while parsing value containing decimal

+++ src/main/java/com/geofriend/expressionevaluator/Expression.java
@@ -256,7 +256,7 @@ public class Expression {
         }
 
         public int asInt() {
-            return Integer.parseInt(result);
+            return (int)Double.parseDouble(result);
         }
 
         public double asDouble() {
